update files under project directory
