+++ project/model/__init__.py
@@ -1,3 +1,1 @@
 from .fitter import Fitter
-from .rpn_fitter import RPNFitter
-from .vit_fitter import ViTFitter
