+++ project/training/split.py
@@ -1,5 +1,5 @@
 from sklearn.model_selection import train_test_split
-from dataset.dataset import VALDODataset
+from project.dataset.valdo import VALDODataset
 
 def split_train_val_datasets(df, transform, test_size=0.2, random_state=42):
     train_df, val_df = train_test_split(
